refactor: break host_service_registry.py into manageable modules
- Extract protocol validation to protocol_validation.py (61 lines)
- Extract storage/cleanup to registry_storage.py (63 lines)
- Extract registration operations to registry_operations.py (223 lines)
- Extract query operations to registry_queries.py (249 lines)
- Extract diagnostics to registry_diagnostics.py (137 lines)
- Main file now composes mixins (71 lines, down from 688)
- Maintain backwards compatibility with re-exports
- All 104 related tests passing
